Refactor: move methods to network.cpp
