+++ command.cpp
@@ -24,6 +24,7 @@
 #include "map_cell.hpp"
 #include "mef.hpp"
 #include "menu.hpp"
+#include "network.hpp"
 #include "quest.hpp"
 #include "shop.hpp"
 #include "snail/application.hpp"
