+++ turn_sequence.cpp
@@ -22,6 +22,7 @@
 #include "map_cell.hpp"
 #include "mef.hpp"
 #include "menu.hpp"
+#include "network.hpp"
 #include "quest.hpp"
 #include "ui.hpp"
 #include "variables.hpp"
