+++ variables.hpp
@@ -182,6 +182,9 @@ ELONA_EXTERN(int dirsub);
 ELONA_EXTERN(int mousex);
 ELONA_EXTERN(int mousey);
 
+// network.cpp
+ELONA_EXTERN(elona_vector1<std::string> netbuf);
+
 ELONA_EXTERN(elona_vector1<int> _randcolor);
 ELONA_EXTERN(elona_vector1<int> aniref);
 ELONA_EXTERN(elona_vector1<int> anix);
@@ -349,7 +352,6 @@ ELONA_EXTERN(elona_vector1<std::string> msg);
 ELONA_EXTERN(elona_vector1<std::string> msgtemp);
 ELONA_EXTERN(elona_vector1<std::string> msgtempprev);
 ELONA_EXTERN(elona_vector1<std::string> musicfile);
-ELONA_EXTERN(elona_vector1<std::string> netbuf);
 ELONA_EXTERN(elona_vector1<std::string> newsbuff);
 ELONA_EXTERN(elona_vector1<std::string> playerheader);
 ELONA_EXTERN(elona_vector1<std::string> q);
@@ -746,7 +748,6 @@ int imeget();
 int key_direction();
 int access_item_db(int);
 int get_required_craft_materials();
-int voting_box();
 int try_to_cast_spell();
 int try_to_reveal();
 int can_evade_trap();
@@ -809,10 +810,6 @@ int map_placeupstairs(int = 0, int = 0);
 int map_trap(int = 0, int = 0, int = 0, int = 0);
 int map_web(int = 0, int = 0, int = 0);
 int mapitemfind(int = 0, int = 0, int = 0);
-int net_dl(const std::string&, const std::string&);
-int net_dllist(const std::string&, int = 0);
-int net_read(int = 0);
-int net_send(const std::string&, int = 0);
 int pileremain();
 int prodcheck();
 int putcard(int = 0, int = 0);
@@ -982,8 +979,6 @@ void initialize_craft_material_data();
 void get_random_item_id();
 void set_item_info();
 void end_dmghp();
-void initialize_server_info();
-void show_chat_dialog();
 void init_character_skills(int);
 void get_hungry(int);
 void proc_turn_end(int);
@@ -1218,7 +1213,6 @@ void modrank(int = 0, int = 0, int = 0);
 void msg_clear();
 void msg_halt();
 void msg_newline();
-void netload(const std::string&);
 void opencard(int = 0);
 void page_load();
 void page_save();
